requirements.txt
flask
flask-cors
transformers
sentence-transformers
torch
faiss-cpu
numpy
pickle5
requests
tqdm